refactor: pr review. add assert to ae def, controllers for stow, services for controllers
