+++ packages/static-wado-util/lib/assertions/index.js
@@ -1,0 +1,7 @@
+const assertAeDefinition = require("./assertAeDefinition.js");
+
+const assertions = {
+  assertAeDefinition,
+};
+
+module.exports = assertions;
