+++ packages/static-wado-util/lib/index.js
@@ -7,6 +7,7 @@ exports.qidoFilter = require("./qidoFilter");
 exports.loadConfiguration = require("./loadConfiguration");
 exports.aeConfig = require("./aeConfig");
 exports.staticWadoConfig = require("./staticWadoConfig");
+exports.assertions = require("./assertions");
 
 exports.configureProgram = configureProgram;
 exports.program = program;
