+++ packages/static-wado-webserver/lib/controllers/client/staticControllers.mjs
@@ -1,12 +1,12 @@
+/* eslint-disable import/prefer-default-export */
 import express from "express";
 import { gzipHeaders } from "../../adapters/responseAdapters.mjs";
 
-export default function staticController(staticFilesDir) {
+export function defaultGetStaticController(staticFilesDir) {
   return express.static(staticFilesDir, {
-    index: "index.json.gz",
+    index: "index.html",
     setHeaders: gzipHeaders,
     extensions: ["gz"],
     redirect: false,
-    fallthrough: true,
   });
 }
