+++ packages/static-wado-webserver/lib/controllers/server/staticControllers.mjs
@@ -1,11 +1,13 @@
+/* eslint-disable import/prefer-default-export */
 import express from "express";
 import { gzipHeaders } from "../../adapters/responseAdapters.mjs";
 
-export default function staticController(staticFilesDir) {
+export function defaultGetStaticController(staticFilesDir) {
   return express.static(staticFilesDir, {
-    index: "index.html",
+    index: "index.json.gz",
     setHeaders: gzipHeaders,
     extensions: ["gz"],
     redirect: false,
+    fallthrough: true,
   });
 }
